make independent cross-docking workflow in docking subpackage
